Fix in-app VNC viewer and rootfs UI state.
Write RFB security type as a single byte, restore rootfs installed flag
when files exist on disk, and use openbox/xterm instead of XFCE.

+++ app/src/main/java/com/proot/cowork/data/vnc/RfbClient.kt
@@ -124,7 +124,7 @@ class RfbClient(
             types.contains(1) -> 1 // None
             else -> types.first()
         }
-        out.writeInt(chosen)
+        out.writeByte(chosen)
         out.flush()
 
         if (chosen == 1) {
